refactor(reviewer-bot): extract shared support modules
