+++ scripts/reviewer_bot_lib/__init__.py
@@ -1,0 +1,1 @@
+"""Shared reviewer-bot logic modules."""
